+++ packages/client/src/pages/GameResult/index.tsx
@@ -1,11 +1,11 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import { citizen, mafia, primaryLight, white, titleActive } from '@constants/colors';
-import { ResultCard } from '@components/Card';
 import { ButtonSizeList, ButtonThemeList, DefaultButton } from '@src/components/Button';
 import { PlayerResult } from '@mafia/domain/types/game';
 import { Link, useLocation, useHistory } from 'react-router-dom';
 import { useUserInfo } from '@src/contexts/userInfo';
+import ResultCardList from '@src/lists/ResultCardList';
 
 interface LocationState {
   playerResultList: PlayerResult[];
@@ -40,11 +40,7 @@ const GameResult = () => {
         </span>
         <span css={personalResultTitleStyle}>YOU {winOrLose}</span>
       </div>
-      <div css={resultCardListStyle}>
-        {playerResultList.map(({ userName, job, result }) => (
-          <ResultCard key={userName} userName={userName} job={job} win={result} />
-        ))}
-      </div>
+      <ResultCardList playerResultList={playerResultList} />
       <div css={buttonDivStyle}>
         <DefaultButton
           size={ButtonSizeList.MEDIUM}
@@ -101,23 +97,6 @@ const teamResultTitleColorStyle = (team: string) => css`
 const personalResultTitleStyle = css`
   font-weight: bold;
   font-size: 80px;
-`;
-
-const resultCardListStyle = css`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-  overflow-y: scroll;
-
-  gap: 20px 30px;
-  width: 100%;
-  height: 100%;
-  padding: 30px 40px;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
 `;
 
 const buttonDivStyle = css`
